+++ pymc3/distributions/multivariate.py
@@ -488,6 +488,16 @@ class Dirichlet(Continuous):
 
     def __init__(self, a, transform=transforms.stick_breaking,
                  *args, **kwargs):
+
+        if not isinstance(a, pm.model.TensorVariable):
+            if not isinstance(a, list) and not isinstance(a, np.ndarray):
+                raise TypeError(
+                    'The vector of concentration parameters (a) must be a python list '
+                    'or numpy array.')
+            a = np.array(a)
+            if (a <= 0).any():
+               raise ValueError("All concentration parameters (a) must be > 0.")
+
         shape = np.atleast_1d(a.shape)[-1]
 
         kwargs.setdefault("shape", shape)